MODULE: Extract interface for GatewaySenderAttributes.
